requirements.txt
Django==1.4.2
PIL==1.1.7
dj-database-url==0.2.1
django-allauth==0.8.2
django-avatar==1.0.5
django-social-auth==0.7.7
 -e git+https://github.com/toastdriven/django-tastypie.git@342076cf6bfacd1b8734e02e6ffe225e590114c4#egg=django_tastypie-dev
httplib2==0.7.6
mimeparse==0.1.3
oauth2==1.5.211
psycopg2==2.4.5
python-dateutil==1.5
python-openid==2.2.5
wsgiref==0.1.2
